refactor(ui): extract ContentWorkbench into modular components
Split `ContentWorkbench.tsx` into smaller, cohesive UI blocks (`WorkbenchHeader`, `SourceContextPane`, `CharlieFeedbackBanner`, `EditorBody`, `AICommandCenter`) and extracted state/logic into `useWorkbenchLogic.ts`. Implemented barrel exports for cleaner imports and separated types. Ensures maintainability while retaining all core physical connectivity.
